requirements.txt
fastapi[standard]
#psycopg2-binary==2.9.6
sqlmodel==0.0.8
uvicorn==0.22.0
aiosqlite
alembic==1.11.1
google-cloud-pubsub
google-cloud-storage